requirements.txt
Flask
beautifulsoup4
pandas
numpy
openpyxl
gunicorn
supabase
python-dotenv
